Add command: rsync-dotfiles
I use this so I can use my favorite dotfiles environment on remote
machines, without me needing to clone the dotfiles repo, copy over git
stuff (not sure if this is implemented). Also, since my dotfiles is now
also dependend on Haskell and the Shelly library (sigh), it's not
feasible to expect it to be installed on all small machines I SSH to.