Gõ tiếng Việt: Make-Ibus-work-in-fx helpers
Previously I wasn't able to type Vietnamese in firefox and had an
annoying "type in terminal then copy+paste" dance each time I wanted to
write a message.

This is a solution in the way that the problem will be "fixed" after a
win+b click on the keyboard. More than good enough.