Add all the plugins I still care about to vim/pack
I add them as just files, no clever package management. The reason is that I
would actually like to use something like git submodules, but it adds one level
of "thinking" when I do things like creating this dotfile on a new computer,
or when rsyncing the files to somewhere.